Add ChangePasswordCoordinator
