+++ Projects/App/Sources/Present/Auth/FindPassword/Password/ChangePasswordViewController.swift
@@ -1,9 +1,5 @@
-//
-//  ChangePasswordViewController.swift
-//  Choice
-//
-//  Created by 민도현 on 2023/06/08.
-//  Copyright © 2023 dohyeon. All rights reserved.
-//
+import UIKit
 
-import Foundation
+class ChangePasswordViewController: BaseVC<ChangePasswordViewModel> {
+    
+}
