+++ Projects/App/Sources/Present/Auth/FindPassword/Password/ChangePasswordViewModel.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+class ChangePasswordViewModel: BaseViewModel {
+    
+}
